+++ tests/test_transforms/test_decomposition/test_base_change_points.py
@@ -4,7 +4,11 @@ import pytest
 from ruptures import Binseg
 
 from etna.datasets import TSDataset
+from etna.datasets import generate_ar_df
+from etna.transforms.decomposition.base_change_points import BaseChangePointsModelAdapter
 from etna.transforms.decomposition.base_change_points import RupturesChangePointsModel
+
+N_BKPS = 5
 
 
 @pytest.fixture
@@ -19,8 +23,15 @@ def df_with_nans() -> pd.DataFrame:
     return df["segment_1"]
 
 
+@pytest.fixture
+def simple_ar_ts(random_seed):
+    df = generate_ar_df(periods=125, start_time="2021-05-20", n_segments=1, ar_coef=[2], freq="D")
+    df_ts_format = TSDataset.to_dataset(df)
+    return TSDataset(df_ts_format, freq="D")
+
+
 def test_fit_transform_with_nans_in_middle_raise_error(df_with_nans):
-    change_point_model = RupturesChangePointsModel(Binseg(), n_bkps=5)
+    change_point_model = RupturesChangePointsModel(change_point_model=Binseg(), n_bkps=N_BKPS)
     with pytest.raises(ValueError, match="The input column contains NaNs in the middle of the series!"):
         _ = change_point_model.get_change_points_intervals(df=df_with_nans, in_column="target")
 
@@ -34,9 +45,29 @@ def test_build_intervals():
         (pd.Timestamp("2020-01-18"), pd.Timestamp("2020-02-24")),
         (pd.Timestamp("2020-02-24"), pd.Timestamp.max),
     ]
-    intervals = RupturesChangePointsModel._build_intervals(change_points=change_points)
+    intervals = BaseChangePointsModelAdapter._build_intervals(change_points=change_points)
     assert isinstance(intervals, list)
     assert len(intervals) == 4
     for (exp_left, exp_right), (real_left, real_right) in zip(expected_intervals, intervals):
         assert exp_left == real_left
         assert exp_right == real_right
+
+
+def test_get_change_points_intervals_format(simple_ar_ts):
+    change_point_model = RupturesChangePointsModel(change_point_model=Binseg(), n_bkps=N_BKPS)
+    intervals = change_point_model.get_change_points_intervals(
+        df=simple_ar_ts.to_pandas()["segment_0"], in_column="target"
+    )
+    assert isinstance(intervals, list)
+    assert len(intervals) == N_BKPS + 1
+    for interval in intervals:
+        assert len(interval) == 2
+
+
+def test_get_change_points_format(simple_ar_ts):
+    change_point_model = RupturesChangePointsModel(change_point_model=Binseg(), n_bkps=N_BKPS)
+    intervals = change_point_model.get_change_points(df=simple_ar_ts.to_pandas()["segment_0"], in_column="target")
+    assert isinstance(intervals, list)
+    assert len(intervals) == N_BKPS
+    for interval in intervals:
+        assert isinstance(interval, pd.Timestamp)